Copy test sources, don't mount them.
Instead of using docker-compose "volumes" to mount the test service
sources into the test service containers, COPY them in the Dockerfile.

This makes for marginally slower development, since the last layers of
the image need to be rebuild every time the service source is changed.

The benefit is that we can use remote docker in CI.  Remote docker
doesn't seem to do volumes.