Use dd-trace-cpp instead of dd-opentracing-cpp (#23)
This revision removes the dependence on [dd-opentracing-cpp][1] and on [opentracing-cpp][2]. Instead, the newer [dd-trace-cpp][3] is used for tracing.

In addition to changing the implementation detail that is the underlying tracing library, I also changed how the module is configured. Previously, users could specify tracer configuration via a `datadog` block directive containing JSON, e.g.
```nginx
http {
    datadog {
        "agent_url": "http://dd-agent:8126",
        "sample_rate": 0.1,
        "propagation_style_extract": ["B3", "Datadog"]
    }
    ...
}
```

Now, the most consequential configuration options (agent URL, propagation styles, and sample rate) have their own configuration directives, and the `datadog { ... }` block directive has been removed. The example above now looks like:
```nginx
http {
    datadog_agent_url http://dd-agent:8126;
    datadog_sample_rate 0.1;
    datadog_propagation_styles B3 Datadog;
    ...
}
```

See the changes to `doc/API.md` for a description of the new configuration directives. The `datadog_sample_rate` directive is particularly sophisticated and is designed to obviate the need for specifying sampling rules.

Environment variables are still supported.

Another change is that `datadog_*` directives are no longer allowed in `if` contexts. The directive that I guessed was most likely to be used in an `if` context, `datadog_tag`, does not work reliably in `if` contexts. I tested it and decided it was not worth making work. This we might need to revisit per user feedback.

[1]: https://github.com/DataDog/dd-opentracing-cpp
[2]: https://github.com/opentracing/opentracing-cpp
[3]: https://github.com/DataDog/dd-trace-cpp