remove "example" from CI checks
Previously, I added a step to the CircleCI checks that runs the
example/ using the most recent release. At one point, I restructured
the repository and the example/ stopped working, but I didn't notice.
The CI check was an attempt to prevent that in the future.

Now it's a pain, though, because anything that might "break" the example
cannot be "fixed" in CI without a release.

Let's revisit how to automate verification that example/ is working.
For now, I'm removing the check to reduce operational burden.