Fix test environment for CI.
CI jobs were failing due to the dd-trace NPM package now requiring a
newer version of Node than is available on alpine:3.15. So, I upgraded
all references of alpine:3.15 to the newer alpine:3.19.

Then, locally, the uWSGI test image was failing to build due to pip3's
"externally-managed-environment" error. So, I passed the flag
"--break-system-packages" to "pip3 install" in the relevant dockerfiles.