Fix race condition
After submission of the final WAF run in the header filter, the headers
can still be modified by subsequent header filters. If the headers are
read from the thread serializaing data for the WAF after ngx_list_push,
but before the new element is written to, unininitalized data is read.

Additionally, nginx doesn't seem to guarantee that tasks can see memory
written to before the task is submitted. Establish such guarantee
through release-acquire for the final WAF run. This was already being
done for the initial waf run.