fix: bypass context propagation if headers are not initialized (#213)
When processing HTTP/0.9 requests, the HTTP headers data structure
is not initialized because this version of HTTP does not support headers.
This leads to a crash when the code attempts to add context propagation
information to an uninitialized list of headers.

We have implemented a check to bypass context propagation if the headers
are not initialized.

Resolves #212

Changes:
  - move headers logic into `common` folder.
  - add integration test with HTTP/0.9.